more merge conflict fixes
