+++ sdk/highlight-run/src/client/__tests__/index.test.tsx
@@ -1,5 +1,5 @@
 import { Highlight } from '../index'
-import { LDClientMin } from '../../integrations/launchdarkly'
+import { LDClientMin } from '../../integrations/launchdarkly/types/LDClient'
 import { vi } from 'vitest'
 
 describe('LD integration', () => {
@@ -18,6 +18,11 @@ describe('LD integration', () => {
 	})
 
 	it('should handle register', () => {
+		const worker = (globalThis.Worker as unknown as typeof Worker).prototype
+		worker.postMessage = vi.fn(
+			(_message: unknown, _options?: unknown) => null,
+		)
+
 		const client: LDClientMin = {
 			track: vi.fn(),
 			identify: vi.fn(),
@@ -28,11 +33,13 @@ describe('LD integration', () => {
 		expect(client.addHook).not.toHaveBeenCalled()
 		expect(client.identify).not.toHaveBeenCalled()
 		expect(client.track).not.toHaveBeenCalled()
+		expect(worker.postMessage).not.toHaveBeenCalled()
 
 		highlight.identify('123', {})
 		highlight.addProperties('test', {})
 		// noop for launchdarkly
 		expect(client.identify).not.toHaveBeenCalled()
 		expect(client.track).toHaveBeenCalled()
+		expect(worker.postMessage).toHaveBeenCalled()
 	})
 })
