+++ sdk/highlight-run/src/integrations/launchdarkly/types/LDClient.ts
@@ -1,0 +1,7 @@
+import { Hook } from './Hooks'
+
+export interface LDClientMin {
+	track(key: string, data?: any, metricValue?: number): void
+	identify(ctx: any): void
+	addHook(hook: Hook): void
+}
